Implement basics for loading MCD files
The FileProcessor is now an abstract class with McdProcessor and
MsrdProcessor responsible for the respective file formats. The MCD files
are no longer loaded with the generated Kaitai classes, but with the
MCStream library. File verification and meta data printing is
implemented. What is left to do is to do the actual conversion.